# docs/swagger.yaml
basePath: /v1
definitions:
...
  entity.User:
    properties:
      admin:
        type: boolean
      bio:
        type: string
      comments_count:
        type: integer
      confirmed:
        type: boolean
      email:
        type: string
      followers:
        items:
          type: string
        type: array
      followers_count:
        type: integer
      following:
        items:
          type: string
        type: array
      following_count:
        type: integer
      has_avatar:
        type: boolean
      last_seen:
        type: integer
      likes:
        items:
          type: string
        type: array
      likes_count:
        type: integer
      location:
        type: string
      member_since:
        type: integer
      name:
        type: string
      password:
        type: string
      submissions_count:
        type: integer
      type:
        type: string
      url:
        type: string
      username:
        type: string
    type: object
  errors.ErrorResponse:
    properties:
      details: {}
      message:
        type: string
      status:
        type: integer
    type: object
  pagination.Pages:
    properties:
      items: {}
      page:
        type: integer
      page_count:
        type: integer
      per_page:
        type: integer
      total_count:
        type: integer
    type: object
...
host: api.saferwall.com
info:
  contact:
    email: support@saferwall.com
    name: API Support
    url: https://about.saferwall.com/contact.html
  description: Interact with Saferwall Malware Analysis Platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://about.saferwall.com/tos
  title: Saferwall Web API
  version: "1.0"
paths:
...
  /users/:
    get:
      consumes:
      - application/json
      description: List users.
      parameters:
      - description: Number of items per page
        in: query
        name: per_page
        type: integer
      - description: Specify the page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/pagination.Pages'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/entity.User'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Retrieves a paginated list of users
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create a new user.
      parameters:
      - description: User data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Create a new user
      tags:
      - User
  /users/{username}:
    get:
      consumes:
      - application/json
      description: Retrieves information about a user.
      parameters:
      - description: User ID
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get user information by user ID
      tags:
      - User
  /users/{username}/:
    delete:
      description: Deletes a user by ID.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/entity.User'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Deletes a user
      tags:
      - User
    patch:
      consumes:
      - application/json
      description: Replace a user document with a new user's document.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: New user data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Update a user object (full update)
      tags:
      - User
  /users/{username}/avatar/:
    post:
      consumes:
      - application/json
      description: Change user avatar
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: User data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.UpdateEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Update user avatar
      tags:
      - User
  /users/{username}/comments/:
    get:
      consumes:
      - application/json
      description: List of comments by a user.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Number of items per page
        in: query
        name: per_page
        type: integer
      - description: Specify the page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pagination.Pages'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Returns a paginated list of a user's comments
      tags:
      - User
  /users/{username}/email/:
    post:
      consumes:
      - application/json
      description: Change email for logged-in users.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: User data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.UpdateEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Update email for authenticated users
      tags:
      - User
  /users/{username}/follow/:
    post:
      consumes:
      - application/json
      description: Start following a user.
      parameters:
      - description: Target user to follow
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Follow a user
      tags:
      - User
  /users/{username}/followers/:
    get:
      consumes:
      - application/json
      description: List of users who follow a user.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Number of items per page
        in: query
        name: per_page
        type: integer
      - description: Specify the page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pagination.Pages'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Returns a paginated list of a user's followers
      tags:
      - User
  /users/{username}/following/:
    get:
      consumes:
      - application/json
      description: List of users a user follows.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Number of items per page
        in: query
        name: per_page
        type: integer
      - description: Specify the page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pagination.Pages'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Returns a paginated list of a user's following
      tags:
      - User
  /users/{username}/likes/:
    get:
      consumes:
      - application/json
      description: List of likes of a user.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Number of items per page
        in: query
        name: per_page
        type: integer
      - description: Specify the page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pagination.Pages'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Returns a paginated list of a user's likes
      tags:
      - User
  /users/{username}/password/:
    post:
      consumes:
      - application/json
      description: Change password for logged-in users.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: User data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Update password for authenticated users
      tags:
      - User
...
  /users/{username}/unfollow/:
    post:
      consumes:
      - application/json
      description: Stop following a user.
      parameters:
      - description: Target user to unfollow
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Unfollow a user
      tags:
      - User
...
schemes:
- https
securityDefinitions:
  Bearer:
    description: 'Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"